ReceiptTrac/requirements.txt
flask==3.0.0
openai==1.12.0
google-cloud-vision==3.7.0
Pillow==10.2.0
python-dotenv==1.0.0
pandas==2.2.0
requests==2.31.0
Werkzeug==3.0.1
gunicorn==21.2.0

# Barcode scanning
pyzbar==0.1.9

# PDF Report generation
reportlab==4.0.9

# Additional utilities
numpy==1.26.4
python-dateutil==2.8.2
